arm64: dma-mapping: Attach IOMMUs as groups
Some devices behind IOMMUs might end up in IOMMU groups due to DMA
aliasing.  The iommu_attach_device API expects a single device in a
group all by its lonesome, so it fails when a device is passed in that's
part of a group.  Use iommu_attach_group to accommodate these types of
devices.

Change-Id: I9601efd300c7f428c4576e6ecf6d31791d0b47bd
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>