scsi: ufs: read device descriptor once
We are reading the device descriptor once for applying the device
quirks and then device descriptor isn't saved anywhere for future
use. Now that we also need the bDeviceSubClass parameter (of device
descriptor), this change refactors the code to read the device
descriptor once and save the imprortant parameters for later use
by driver.

Change-Id: Icb5d0380a9dcac0e7bd042d9be3abd8bcf6ead48
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>