drivers: cpuidle: lpm-levels: Do not set affinity level for 0 mode id
A PSCI ID of zero doesn't require any setting from the firmware. In this
case, incrementing the affinity level would be incorrect, especially for
a single core system where the CPU and Cluster states are linearized.
The Sleep driver in the kernel needs to different between CPU/Cluster
states to send System votes through before sleep.

Do not increment affinity level when PSCI ID is zero.

Change-Id: I4739ee1ac893d30e5ab5f24b1c388ef58acda06a
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>