ANDROID: sched/fair: if sync flag ignored, try to place in same cluster
If the sync flag is ignored because the current cpu is not in
the affinity mask for the target of a sync wakeup (usually
binder call), prefer to place in the same cluster if
possible. The main case is a "top-app" task waking a
"foreground" task when the top-app task is running on
a CPU that is not in the foreground cpuset. This patch
causes the "top-app" search order be used (big cpus first)
even if schedtune.boost is 0 when the sync flag failed.

Bug: 110464019
Change-Id: Ib026231cee5c8b9d616d4ad4b92dd8502e79d0c0
Signed-off-by: Todd Kjos <tkjos@google.com>
Keep changes from 9f7516a28d7ca45d2d69352d70f4df4365041dd4.
Signed-off-by: idkwhoiam322 <snkm1999@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>