binder: add function to copy binder object from buffer
When creating or tearing down a transaction, the binder driver
examines objects in the buffer and takes appropriate action.
To do this without needing to dereference pointers into the
buffer, the local copies of the objects are needed. This patch
introduces a function to validate and copy binder objects
from the buffer to a local structure.

Change-Id: I8069fb6d62ed9ab2a359936a41436a9caf79c2c8
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>